Another attempt at travis bundler caching.
For rspec/rspec-dev#101. The changes in #102 don’t seem
to be caching properly. I’m switching back to a single
shared bundle directory but manually caching it rather
than using `cache: bundler` so that `bundle clean` doesn’t
run.

I plan to put this in rspec-dev (and all repos) if it works
but wanted to test on just one repo for now.