Correct inaccurate comment.
It’s actually not quite the hot spot it says it here.

Encoded strings are only created when an expectation fails
that uses a diffable matcher.  I believe that expectations normally
pass (since people usually try to keep their test suite green
and only have a small number of failing specs at a time), so encoded
strings are not created by every expectation.