FuzzyMatcher is friendly to custom classes subclassed from Array
When comparing subclasses of Array, FuzzyMatcher would call #to_a on one object
but not both, which could cause the comparison to fail if #to_a was overridden.
Additionally, some subclasses of Array may define their own idea of 'equality',
which should be respected by FuzzyMatcher.

On the other hand, if the 'expected' value is an Array, it is reasonable to
coerce the subclass of Array to an actual Array. Since Enumerable does not
define #to_ary, but only #to_a, use #to_a for this purpose.