Reduce memory usage by using frozen string literals
I was running memory_profiler on a spec with a large number of examples
(950), and it surfaced several places in rspec-support that aren't using
frozen strings.

Originally, it was `::` in lib/rspec/support/recursive_const_methods.rb
that caught my attention, but I figure it can be enabled everywhere.

I did not enable it for specs, because I figure that wouldn't have as
much impact since it would only affect rspec-support's specs.