Update Differ to handle arrays of single-line strings
(instead of skipping them). This behavior was originally added in in
rspec-expectations commit 044b0a6ff1, as part of
https://github.com/rspec/rspec-expectations/issues/115, which was
intended to `expect("a").to eq("b")` from showing an extra diff, when
the error message already showed the whole diff (so the diff itself was
kind of unnecessary). There are tests already enforcing _that_, so we
don't need to add any, but this change retains that behavior (though
single-element single-line array comparisons do now get a diff).