Override RSpec::Support::StdErrSplitter#clone
With the default definition, cloning a StdErrSplitter results in the
clone having the same @orig_stderr instance as the original Splitter
has, which means that reopening either Splitter affects both (since the
underlying stream is not cloned).

This is different from the behavior of an actual stream, in a way that
causes a problem for the CaptureStreamToTempfile in rspec-matchers (used
by the `output.to_stderr_from_any_process` matcher), because it clones
the original stream to store it for restoration, and then `reopen`s it
at a Tempfile temporarily. But then when it goes back to _restore_ the
stream, it doesn't actually do so, since the clone's stream reference is
the same instance as the original Splitter has - it just reopens it with
the Tempfile again.